requirements.txt
Flask==2.1.1
cassandra-driver==3.25.0
Werkzeug==2.0.1
# Add other dependencies here
